refactor(auth): one useAuth, not two
Card 12's remaining half. With the unsound `NonNullable` cast removed in wave 1,
`hooks/useAuth` and `context/hooks/useAuth` were the same eight lines over the
same context, differing only in which export style they offered. Both were live:
19 modules imported one, 7 the other, and the test suite mocked whichever path
its subject happened to use.

`hooks/useAuth` survives (it offers both the default and the named export);
`context/hooks/useAuth` is deleted and its 19 importers plus 13 test mocks now
name the surviving module.
