+++ src/components/MessageInput/index.tsx
@@ -1,7 +1,7 @@
 import { useState, FormEvent, ChangeEvent, KeyboardEvent } from 'react';
 import { IconButton, InputAdornment, TextField } from '@mui/material';
 import SendIcon from '@mui/icons-material/Send';
-import useAuth from '@/context/hooks/useAuth';
+import useAuth from '@/hooks/useAuth';
 import { sendMessage } from '@/services/firebase/chat';
 import './styles.css';
 import { Trans, useTranslation } from 'react-i18next';
