+++ src/hooks/__tests__/useReturnToStart.test.tsx
@@ -10,7 +10,7 @@ vi.mock('@/services/firebase/chat', () => ({
   sendMessage: vi.fn(),
 }));
 
-vi.mock('@/context/hooks/useAuth', () => ({
+vi.mock('@/hooks/useAuth', () => ({
   default: () => ({
     user: {
       uid: 'test-user',
