+++ src/hooks/useHybridPlayerList.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useLocalPlayers } from './useLocalPlayers';
 import usePlayerList from './usePlayerList';
-import useAuth from '@/context/hooks/useAuth';
+import useAuth from '@/hooks/useAuth';
 import useMessages from '@/context/hooks/useMessages';
 import { orderedMessagesByType } from '@/helpers/messages';
 import { getTurnFields } from '@/helpers/actionTurn';
